security(coordinator-api): wallet trust-boundary hardening and exact monetary semantics (B3/B4)
- Close wallet RPC/secret trust-boundary violations and remove client-supplied
  RPC/private-key paths.
- Replace monetary floats with Decimal/atomic units across coordinator wallet,
  marketplace, staking, settlement, and cross-chain models/services.
- Add money.py helpers for exact atomic<->Decimal conversions with float rejection.
- Update affected Alembic migration (e9cf23ae4640) to migrate remaining Float
  monetary columns to Numeric.
- Add property/regression tests for conversion round-trips, fee calculation, and
  adapter float rejection.

+++ apps/coordinator-api/src/coordinator_api/contexts/staking/services/staking_service.py
@@ -106,8 +106,8 @@ class StakingService:
         user_address: str,
         status: StakeStatus | None = None,
         agent_wallet: str | None = None,
-        min_amount: float | None = None,
-        max_amount: float | None = None,
+        min_amount: Decimal | None = None,
+        max_amount: Decimal | None = None,
         agent_tier: PerformanceTier | None = None,
         auto_compound: bool | None = None,
         page: int = 1,
@@ -280,7 +280,7 @@ class StakingService:
         accuracy: Decimal,
         successful: bool,
         response_time: Decimal | None = None,
-        compute_power: float | None = None,
+        compute_power: Decimal | None = None,
         energy_efficiency: Decimal | None = None,
     ) -> AgentMetrics:
         """Update agent performance metrics"""
